perf: chart animations
